Add test case class for testing helpers
